finally hooked up phone to debug
Now you can shoot while moving. The problem was that I had to separate the touch functions, it was getting the [0] touch on the list which is always the same so it interrupted, all I had to do was actually listen to the second one and made the function fire only when those parameters are valid. [1]